Initial code re-work and swing implementation
Moved the swing target down to create a better swing tether visual.

Added text prompts to stalegtights

Added a player tag to the player blueprint.

Added a stalegtight to the scene due to a bug where I couldnt update the versions in the scene already.

Made a copy of the original player (to be deleted later once all the old features are implemented into the new blueprint.)

Added swinging mechanics/physics to the player.